feat: create companies application
