+++ EDMS/companies/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CompaniesConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "companies"
