+++ EDMS/dashboards/templates/dashboards/dashboard.html
@@ -3,58 +3,7 @@
     Dashboard
 {% endblock title %}
 {% block content %}
-    <!-- Page Wrapper -->
-    <div id="wrapper">
-        {% include "dashboards/sidebar.html" %}
-        <!-- Content Wrapper -->
-        <div id="content-wrapper" class="d-flex flex-column">
-            <!-- Main Content -->
-            <div id="content">
-                {% include "dashboards/topbar.html" %}
-                <!-- Begin Page Content -->
-                {% block main %}
-                {% endblock main %}
-                <!-- End of Page Content -->
-            </div>
-            <!-- End of Main Content -->
-            <!-- Footer -->
-            <footer class="sticky-footer bg-white">
-                <div class="container my-auto">
-                    <div class="copyright text-center my-auto">
-                        <span>Created by Krzysztof Czernicki</span>
-                    </div>
-                </div>
-            </footer>
-            <!-- End of Footer -->
-        </div>
-        <!-- End of Content Wrapper -->
-    </div>
-    <!-- End of Page Wrapper -->
-    <!-- Scroll to Top Button-->
-    <a class="scroll-to-top rounded" href="#page-top">
-        <i class="fas fa-angle-up"></i>
-    </a>
-    <!-- Logout Modal-->
-    <div class="modal fade"
-         id="logoutModal"
-         tabindex="-1"
-         role="dialog"
-         aria-labelledby="exampleModalLabel"
-         aria-hidden="true">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="exampleModalLabel">Ready to Leave?</h5>
-                    <button class="close" type="button" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">×</span>
-                    </button>
-                </div>
-                <div class="modal-body">Select "Logout" below if you are ready to end your current session.</div>
-                <div class="modal-footer">
-                    <button class="btn btn-secondary" type="button" data-dismiss="modal">Cancel</button>
-                    <a class="btn btn-primary" href={% url 'login' %}>Logout</a>
-                </div>
-            </div>
-        </div>
+    <div class="card-header">
+        <p class="font-weight-bold">Tutaj będa rzeczy</p>
     </div>
 {% endblock content %}
